dev_requirements.txt
flake8==3.3.0
pytest==3.0.7
pytest-django==3.1.2
